v2/requirements.txt
cryptography>=41.0.0
docker==7.1.0
fastapi==0.115.0
firebase-admin==6.5.0
flask-cors==4.0.0
flask-socketio==5.3.6
flask==2.3.3
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-cloud-firestore>=2.18.0
google-cloud-logging>=3.11.0
google-cloud-secret-manager>=2.24.0
google-cloud-storage==2.19.0
gunicorn==22.0.0
ipykernel==6.29.5
jupyter==1.0.0
langchain==0.3.25
langgraph==0.4.8
matplotlib==3.10.3
openai==1.75.0
openpyxl==3.1.5
pandas==2.3.0
pdfkit==1.0.0
pydantic==2.9.2
pymssql==2.3.1
PyMySQL==1.1.1
pyyaml>=6.0
python-dotenv==1.0.0
python-multipart
python-socketio==5.10.0
requests>=2.31.0
scipy==1.15.3
seaborn==0.13.2
SQLAlchemy==2.0.36
uvicorn[standard]==0.32.0
weasyprint==66.0
werkzeug==2.3.7